requirements.txt
streamlit==1.28.1
torchvision==0.16.0
diffusers==0.21.4
transformers==4.34.1
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
scipy==1.11.3
accelerate==0.24.0
safetensors==0.4.0
xformers==0.0.22.post7
huggingface-hub==0.17.3
scikit-image==0.21.0
